add basic api reporter
